add find_target_pattern(string)
uses a python.re regex object  to find all occurences of a pattern within a string. The purpose (at this moment) is to only return values when only one unique match is found.  An error will be raised when more than 1 match is found.
returns `None` when nothing is found